flask-mysql/requirements.txt
Python == 3.10.6
pip==23.3.1


click==8.0.1
colorama==0.4.6
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
mysql-connector-python 8.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
tomli==2.0.1
tomlkit==0.12.1
Werkzeug==2.0.1
WTForms==3.0.1